MWPW-195761: inline component-specific controllers into owning components
Yesil's review noted that checkout-controller, placeholder-options-controller,
and search-controller were each 1:1 with a single component (MasOstCheckoutOptions,
MasOstPlaceholderPanel, MasOstSearch). Controllers are meant to encapsulate
behavior reusable across multiple components — a controller specific to one
component is the wrong abstraction and trains the next AI to copy the pattern.

Changes:
- Inline CheckoutController state + methods into ost-checkout-options.js
- Inline PlaceholderOptionsController into ost-placeholder-panel.js
- Inline SearchController into ost-search.js
- Delete OfferFilterController (no consumers — dead code)
- Each component exposes a back-compat getter (e.g. get checkout() { return this })
  so external readers (ost-code-output, ost-live-preview) that previously did
  panel.querySelector('ost-checkout-options').checkout.workflowStep keep working
  without changes
- Port the controller-class unit tests to component-level tests using @open-wc
  fixture() — covers all the same behaviors via the public API
- Re-export WORKFLOW_STEPS, MODAL_TYPES, DEFAULT_MODAL_BY_OFFER from
  ost-checkout-options.js so external code can still import the constants

Test count: 284 (pre) -> 275 (post). The 9 net-lost tests were asserting
controller-class plumbing ("registers itself on host", "hostConnected callable")
that no longer applies — all behavioral coverage is preserved at the component
level.
